1.4.24 Make no tasks nub smoother

+++ app/src/main/java/com/q335/r49/squaredays/MainActivity.java
@@ -478,13 +478,16 @@ public class MainActivity extends AppCompatActivity implements TasksFrag.OnFragm
                             @Override
                             public void onClick(DialogInterface dialog, int which) {
                                 File logFile = new File(context.getFilesDir(), LOG_FILE);
-                                if (logFile.delete()) {
-                                    pushTask(logEntry.newClearMess());
-                                    setPermABState(null);
-                                    if (GF.isVisible())
-                                        popTasks();
+                                if (logFile.exists()) {
+                                    if (logFile.delete()) {
+                                        pushTask(logEntry.newClearMess());
+                                        setPermABState(null);
+                                        if (GF.isVisible())
+                                            popTasks();
+                                    } else
+                                        Log.d("SquareDays", "Log clear failed!");
                                 } else
-                                    Log.d("SquareDays","Log clear failed!");
+                                    pushTask(logEntry.newClearMess());
                             }
                         })
                         .show();
